refactor: move ws to global comp
